Move module `features.analytics.test` to `services.analytics.test`. `FakeAnalyticsService` was not implemented at the correct location.
