+++ features/location/impl/build.gradle.kts
@@ -55,6 +55,6 @@ dependencies {
     testImplementation(libs.test.turbine)
     testImplementation(libs.test.truth)
     testImplementation(projects.libraries.matrix.test)
-    testImplementation(projects.features.analytics.test)
+    testImplementation(projects.services.analytics.test)
     testImplementation(projects.features.messages.test)
 }
