+++ features/preferences/impl/build.gradle.kts
@@ -66,7 +66,7 @@ dependencies {
     testImplementation(projects.features.rageshake.test)
     testImplementation(projects.features.rageshake.impl)
     testImplementation(projects.features.logout.impl)
-    testImplementation(projects.features.analytics.test)
+    testImplementation(projects.services.analytics.test)
     testImplementation(projects.features.analytics.impl)
     testImplementation(projects.tests.testutils)
 }
